requirements.txt
Babel==2.14.0
blinker==1.7.0
cachelib==0.9.0
certifi==2024.2.2
chardet==3.0.4
click==8.1.7
colorama==0.4.6
dnspython==2.6.1
email_validator==2.1.1
Flask==3.0.2
Flask-Admin==1.6.1
flask-babel==4.0.0
Flask-BabelEx==0.9.4
Flask-Caching==2.1.0
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
googletrans==3.1.0a0
greenlet==3.0.3
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2024.4.1
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
importlib-metadata==7.0.1
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
mysqlclient==2.2.4
PyMySQL==1.1.0
python-dotenv==1.0.1
pytz==2024.1
rfc3986==1.5.0
sniffio==1.3.1
speaklater==1.3
SQLAlchemy==2.0.27
typing_extensions==4.9.0
Werkzeug==3.0.1
WTForms==3.1.2
zipp==3.17.0
